refactor: modernize control flow, regenerate FessConfig, rename api.v2.* keys (#3280)
Mechanical cleanup sweep across the codebase, plus a catch-up regeneration of
the generated config interface and a config-key namespace fix.

- Remove 292 redundant `super();` calls from constructors. 261 of the touched
  files change nothing else.
- Modernize control flow in 87 files: if-chains to switch, colon-switch to
  arrow/expression switch, `instanceof` pattern variables, De Morgan merges,
  redundant casts and `.intValue()` calls dropped, and multi-line log messages
  converted to text blocks.
- Regenerate FessConfig for eight keys that already existed in
  fess_config.properties, and refresh stale javadoc comments. No key,
  accessor, or defaultMap entry is removed.

The regeneration re-declares five accessors that FessProp had been supplying
as tolerant defaults (added in #3162). A generated declaration in FessConfig
plus its SimpleImpl body overrides an interface default, which silently
dropped those fallbacks: a blank configured value returns null and would then
NPE at the call sites that unbox it. The FessProp helpers are therefore
renamed to `...OrDefault` -- a suffix freegen never emits -- and their call
sites follow. FessPropTest pins the blank-value fallback so the same
collision cannot recur unnoticed.

BREAKING: the `api.v2.*` configuration keys are renamed to `api.*`, matching
the rest of the `api.*` namespace. The API version belongs in the endpoint
path, not in a settings key.

  api.v2.param.max.length                    -> api.param.max.length
  api.v2.param.max.array.size                -> api.param.max.array.size
  api.v2.click.max.rt                        -> api.click.max.timestamp
  api.v2.chat.stream.keepalive.interval.ms   -> api.chat.stream.keepalive.interval.ms
  api.v2.chat.rate.limit.per.user.per.minute -> api.chat.rate.limit.per.user.per.minute

These shipped in 15.7.0. An existing override under an old key is no longer
read and the documented default applies instead. The `org.codelibs.fess.api.v2`
package and the `/api/v2` endpoint path are unchanged.

Also restores a `return true;` that had been dropped from the
`/chat/sessions/` arm of CsrfRequirement, which left an empty `if` body.
Behavior was unchanged (the arm fell through to the secure default), but the
explicit rule is back.

+++ src/main/java/org/codelibs/fess/app/service/UserService.java
@@ -255,10 +255,10 @@ public class UserService {
             userBhv.update(entity, op -> {
                 op.setRefreshPolicy(Constants.TRUE);
                 if (seqNo != null && seqNo.longValue() >= 0L) {
-                    op.setIfSeqNo(seqNo.longValue());
+                    op.setIfSeqNo(seqNo);
                 }
                 if (primaryTerm != null && primaryTerm.longValue() >= 0L) {
-                    op.setIfPrimaryTerm(primaryTerm.longValue());
+                    op.setIfPrimaryTerm(primaryTerm);
                 }
             });
             if (logger.isDebugEnabled()) {
